feat: adds try-with-metrics
We often need to track the performance (latency) of I/O bound
features (i.e. HTTP requests, database requests, API response
times), as well as compute bound features (i.e. algorithms).
Knowing the number of times a feature is used (counts), as well
as the volume it receives (gauges) can help us understand which
features are most important to our users, or if a feature is
actively being used, right now.

This adds `tryWithMetrics` to the LogEmitter, to simplify gathering
these kinds of metrics, so all you need to do is supply the writer to
get these metrics into your tracking system.