feat: adds publishing action
This adds a GH Action for publishing to NPM when a release is created in
GH. It also renames the existing action to make it more semantic to its
purpose and updates the NodeJS versions that are tested.

fix: fixes linting errors
fix: removes build step in pr-request action

I think pnpm regressed the --if-present flag. It throws now. I'm just
removing the build step since one doesn't exist. It was there so it was
easier to copy paste this file across libraries.

fix: replaces pull-request code with updated action
fix: changes the nodejs targets for GH Actions