chore: removes unused writer-validator
The the writers are no longer integrated, they are self standing tools
that can be used when subscribing to the log emitter. So there isn't
a call where we would validate the writer anymore. The
formatter-validator is still relevant because the built-in writers
validate the formatter that's passed to them.